:recycle: Move CLI Tests To Seperate Module
